+++ packages/amazonq/src/lsp/chat/messages.ts
@@ -36,9 +36,6 @@ import {
     ShowSaveFileDialogParams,
     LSPErrorCodes,
     tabBarActionRequestType,
-    ShowDocumentParams,
-    ShowDocumentResult,
-    ShowDocumentRequest,
     contextCommandsNotificationType,
     ContextCommandParams,
     openFileDiffNotificationType,
@@ -58,7 +55,7 @@ import { Disposable, LanguageClient, Position, TextDocumentIdentifier } from 'vs
 import * as jose from 'jose'
 import { AmazonQChatViewProvider } from './webviewProvider'
 import { AuthUtil, ReferenceLogViewProvider } from 'aws-core-vscode/codewhisperer'
-import { amazonQDiffScheme, AmazonQPromptSettings, getLogger, messages, openUrl } from 'aws-core-vscode/shared'
+import { amazonQDiffScheme, AmazonQPromptSettings, messages, openUrl } from 'aws-core-vscode/shared'
 import {
     DefaultAmazonQAppInitContext,
     messageDispatcher,
@@ -422,31 +419,6 @@ export function registerMessageListeners(
             targetUri: targetUri.toString(),
         }
     })
-
-    languageClient.onRequest<ShowDocumentParams, ShowDocumentResult>(
-        ShowDocumentRequest.method,
-        async (params: ShowDocumentParams): Promise<ShowDocumentParams | ResponseError<ShowDocumentResult>> => {
-            try {
-                const uri = vscode.Uri.parse(params.uri)
-                if (uri.scheme.startsWith('http')) {
-                    try {
-                        await openUrl(vscode.Uri.parse(params.uri))
-                        return params
-                    } catch (err: any) {
-                        getLogger().error(`Failed to open http from LSP: error: %s`, err)
-                    }
-                }
-                const doc = await vscode.workspace.openTextDocument(uri)
-                await vscode.window.showTextDocument(doc, { preview: false })
-                return params
-            } catch (e) {
-                return new ResponseError(
-                    LSPErrorCodes.RequestFailed,
-                    `Failed to open document: ${(e as Error).message}`
-                )
-            }
-        }
-    )
 
     languageClient.onNotification(contextCommandsNotificationType.method, (params: ContextCommandParams) => {
         void provider.webview?.postMessage({
